# tradequest_spec_v2/Architecture/Detailed_API_Spec.yaml
openapi: 3.0.3
info:
  title: TradeQuest API — Detailed Spec (MVP)
  version: 1.0.0
servers:
  - url: https://api.atqventures.com
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: tq_session
  schemas:
...
    Candle:
      type: object
      required: [ts,o,h,l,c,v]
      properties:
        ts: {type: string, format: date-time}
        o: {type: number}
        h: {type: number}
        l: {type: number}
        c: {type: number}
        v: {type: number}
    Trade:
      type: object
      required: [id,account,venue,symbol,side,qty,avg_price,filled_at]
      properties:
        id: {type: string}
        account: {type: string, example: "spot"}
        venue: {type: string, example: "KRAKEN"}
        symbol: {type: string, example: "BTC/USDT"}
        side: {type: string, enum: ["buy","sell"]}
        qty: {type: number}
        avg_price: {type: number}
        fees: {type: number, default: 0}
        pnl: {type: number, default: 0}
        submitted_at: {type: string, format: date-time, nullable: true}
        filled_at: {type: string, format: date-time}
        order_ref: {type: string, nullable: true}
        session_id: {type: string, nullable: true}
        raw: {type: object, nullable: true}
...
security:
  - cookieAuth: []
paths:
...
  /market/ohlcv:
    get:
      tags: [market]
      summary: Fetch OHLCV
      parameters:
        - {name: symbol, in: query, required: true, schema: {type: string}}
        - {name: tf, in: query, required: true, schema: {type: string, enum: ["1m","5m","15m","1h","4h","1d"]}}
        - {name: limit, in: query, schema: {type: integer, default: 500}}
      responses:
        "200":
          description: Candle[]
          content:
            application/json:
              schema: {type: array, items: {$ref: "#/components/schemas/Candle"}}
  /market/context:
    post:
      tags: [market]
      summary: Candles around a trade
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [trade]
              properties:
                trade: {$ref: "#/components/schemas/Trade"}
                tf: {type: string, default: "1m"}
                pre_mins: {type: integer, default: 30}
                post_mins: {type: integer, default: 60}
      responses:
        "200": {description: Context window}
...